backend/requirements.txt
websockets==10.4
pyaudio
grpcio
grpcio-tools
python-dotenv
